requirement.txt
opencv-python
pillow
matplotlib
torch
torchvision